Spacer周りのメソッド・型をutils/に移動
