+++ src/components/Divider/styled.ts
@@ -1,12 +1,11 @@
 import styled from "styled-components";
-import { SpacerProps, spacer } from "../Spacer/Spacer";
+import { SpacerProps, spacer } from "../../utils/spacerUtils";
 
 type DividerProps = SpacerProps & {
   color: string;
 };
 
 export const Divider = styled.hr<DividerProps>`
-  /* TODO: style.tsに書かれているものもutils/に移動せずに利用して良いのか相談 */
   ${spacer}
   height: 1px;
   border: none;
